+++ packages/components/test/index.js
@@ -19,6 +19,23 @@ import {
   Radio,
   Checkbox,
   Slider,
+  Field,
+  Progress,
+  Donut,
+  Spinner,
+  Avatar,
+  Badge,
+  Close,
+  Alert,
+  Divider,
+  Embed,
+  AspectRatio,
+  AspectImage,
+  Container,
+  NavLink,
+  Message,
+  IconButton,
+  MenuButton,
 } from '../src'
 
 expect.extend(matchers)
@@ -61,6 +78,7 @@ const theme = {
     nav: {
       textDecoration: 'none',
       color: 'inherit',
+      fontWeight: 'normal',
     },
   },
 }
@@ -202,8 +220,18 @@ describe('Grid', () => {
     expect(json).toMatchSnapshot()
   })
 
+  test('renders with mixed columns prop', () => {
+    const json = renderJSON(<Grid columns="1fr 2fr" />)
+    expect(json).toMatchSnapshot()
+  })
+
   test('renders with responsive columns prop', () => {
     const json = renderJSON(<Grid columns={[2, 3, 4]} />)
+    expect(json).toMatchSnapshot()
+  })
+
+  test('renders with mixed columns prop', () => {
+    const json = renderJSON(<Grid columns={[null, '1fr 2fr']} />)
     expect(json).toMatchSnapshot()
   })
 })
@@ -287,10 +315,7 @@ describe('Select', () => {
   test('renders with style props', () => {
     const json = renderJSON(
       <ThemeProvider theme={theme}>
-        <Select
-          mb={3}
-          value='hello'
-        />
+        <Select mb={3} value="hello" />
       </ThemeProvider>
     )
     expect(json).toMatchSnapshot()
@@ -335,6 +360,193 @@ describe('Slider', () => {
     const json = renderJSON(
       <ThemeProvider theme={theme}>
         <Slider />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Field', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Field />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Progress', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Progress />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Donut', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Donut />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Spinner', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Spinner />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Avatar', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Avatar />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Badge', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Badge />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Close', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Close />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Alert', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Alert />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Divider', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Divider />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Embed', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Embed />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('AspectRatio', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <AspectRatio />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('AspectImage', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <AspectImage />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Container', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Container />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('NavLink', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <NavLink />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('Message', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Message />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('IconButton', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <IconButton />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
+})
+
+describe('MenuButton', () => {
+  test('renders', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <MenuButton />
       </ThemeProvider>
     )
     expect(json).toMatchSnapshot()