+++ packages/docs/src/components/code.js
@@ -96,6 +96,7 @@ export const LiveCode = ({ children, preview, xray }) => {
             fontSize: 0,
             color: 'secondary',
             bg: 'highlight',
+            overflow: 'auto',
           }}
         />
       </div>