+++ packages/docs/src/index.js
@@ -3,11 +3,7 @@ import { jsx } from 'theme-ui'
 import Layout from './components/layout'
 
 export const wrapPageElement = ({ element, props }) => (
-  <Layout
-    {...props}
-    fullwidth={props.location.pathname === '/'}
-    children={element}
-  />
+  <Layout {...props} children={element} />
 )
 
 export { default as Banner } from './components/banner'