+++ packages/theme-ui/src/color-modes.js
@@ -5,10 +5,12 @@ import { useThemeUI } from './context'
 import { createColorStyles } from './custom-properties'
 
 const STORAGE_KEY = 'theme-ui-color-mode'
+const HAS_STORAGE = typeof Storage !== 'undefined'
 
 const storage = {
-  get: init => window.localStorage.getItem(STORAGE_KEY) || init,
-  set: value => window.localStorage.setItem(STORAGE_KEY, value),
+  get: init =>
+    (HAS_STORAGE && window.localStorage.getItem(STORAGE_KEY)) || init,
+  set: value => HAS_STORAGE && window.localStorage.setItem(STORAGE_KEY, value),
 }
 
 export const getMediaQuery = () => {